feature: 分页加载

+++ src/main/java/com/gohead/core/dao/ArticleMapper.java
@@ -34,7 +34,7 @@ public interface ArticleMapper {
      * @param map
      * @return
      */
-    public Long getTotalArticles(Map<String, Object> map);
+    public int getTotalArticles(Map<String, Object> map);
 
     /**
      * 添加文章
